Renamed the APICard model to Card
