+++ src/app/card-collection-overview/card-collection-overview/components/card-collection-overview.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 import { CollecteDBService } from '../../../shared/collecteDB/collecte-db.service';
 import { MessageService } from '../../../shared/messages/services/messages.service';
-import { APICard, CardVersion } from '../../../shared/models/api';
+import { Card, CardVersion } from '../../../shared/models/card';
 
 @Component({
   selector: 'app-card-collection-overview',
@@ -10,7 +10,7 @@ import { APICard, CardVersion } from '../../../shared/models/api';
   styleUrls: ['../pages/card-collection-overview.component.scss']
 })
 export class CardCollectionOverviewComponent {
-  cards : APICard[] = [];
+  cards : Card[] = [];
 
   constructor(private collecteDBService: CollecteDBService, private messageService: MessageService) {}
 
@@ -22,7 +22,7 @@ export class CardCollectionOverviewComponent {
     this.collecteDBService.getAllCards().subscribe(fetched => this.cards = fetched);
   }
 
-  countCards(card: APICard) : number {
+  countCards(card: Card) : number {
     var cnt = 0;
     for (var version of card.versions)
       cnt += version.card_count;
